重構戰鬥相關結構，將整數類型改為無符號字符以節省內存，移除未使用的函數，改善代碼可讀性

+++ src/video.h
@@ -32,10 +32,10 @@ typedef struct PAL_Surface {
 } PAL_Surface;
 
 typedef struct PAL_Rect {
-    unsigned short x;
-    unsigned short y;
-    unsigned short w;
-    unsigned short h;
+    int x;
+    int y;
+    int w;
+    int h;
 } PAL_Rect;
 
 extern PAL_Surface *gpScreen;
